% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bottles.R
\docType{data}
\name{bottles}
\alias{bottles}
\title{Bottles data on filling amount of six bottles over five time points.}
\format{
A data frame with 30 rows and 3 variables:
\describe{
  \item{weight}{the response, the weight of material filled into each bottle.}
  \item{time}{the time point of each filling.}
  \item{heads}{the head of the machine that fills the bottle.}
}
}
\usage{
data(bottles)
}
\description{
Ott and Snee (1973) analyzed the filled weight (weight) of bottles filled by
six machine heads (heads) over five time points (time).
}
\references{
\insertRef{bottles}{slgf}
}
\keyword{datasets}
