% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupings.R
\name{groupings}
\alias{groupings}
\title{Groupings finder for two-way layouts.}
\usage{
groupings(data_matrix)
}
\arguments{
\item{data_matrix}{an r by c data matrix.}
}
\value{
\code{groupings} returns the unique possible row-wise groupings of the input two-way layout.
}
\description{
\code{groupings} Computes the possible grouping schemes for a two-way layout with r rows and c columns.
}
\examples{
# Determine the possible row-wise groupings for an 8 by 5 matrix.
groupings(matrix(NA, nrow=8, ncol=4))

}
