% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labeler.R, R/utils.R
\name{labeler}
\alias{labeler}
\title{Labels observations according to group membership.}
\usage{
labeler(nrows, ncols, combo.iteration)

labeler(nrows, ncols, combo.iteration)
}
\arguments{
\item{nrows}{the number of rows in the data matrix.}

\item{ncols}{the number of columns in the data matrix.}

\item{combo.iteration}{the index of the grouping scheme under consideration.}
}
\value{
\code{labeler} returns a vector of 1s and 0s corresponding to the input vector's group membership by index.
}
\description{
\code{labeler} Returns a Boolean indicator for each observation's group membership for a two-way layout.
}
