% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lymphoma.R
\docType{data}
\name{lymphoma}
\alias{lymphoma}
\title{Lymphoma data on genomic hybridization signal from six dogs with normal and tumor
tissue samples taken.}
\format{
A data frame with 6 rows and 2 variables:
\describe{
  \item{V1}{the signals from the normal tissue samples.}
  \item{V2}{the signals from the tumor tissue samples.}
}
}
\usage{
data(lymphoma)
}
\description{
Franck et. al. (2013) analyzes the genomic hybridization signal
measured from normal and tumor tissue samples taken from six dogs.
}
\references{
\insertRef{franckdogs}{slgf}
}
\keyword{datasets}
