% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maketall.R
\name{maketall}
\alias{maketall}
\title{Converts a two-way layout into tall format with row and column index labels.}
\usage{
maketall(data_matrix)
}
\arguments{
\item{data_matrix}{an r by c data matrix.}
}
\value{
\code{maketall} returns a data frame containing the original observations, row labels, and column labels.
}
\description{
\code{maketall} Converts a two-way layout into tall format with row and column index labels.
}
\examples{
library(slgf)
data(lymphoma)
maketall(lymphoma)

}
