% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roadwear.R
\docType{data}
\name{roadwear}
\alias{roadwear}
\title{Roadwear data on four tires, each comprising three compounds,
from a balanced incompleted block design.}
\format{
A data frame with 12 rows and 3 variables:
\describe{
  \item{abrasion}{the measurement of abrasion.}
  \item{compound}{the compound from which each measurement was taken,
  either A, B, C, or D.}
  \item{tire}{the tire from which each measurement was taken,
  either 1, 2, 3, or 4.}
}
}
\usage{
data(roadwear)
}
\description{
Davies (1954) analyzes the wear on four tires, where each tire comprises
three distinct compounds.
}
\references{
\insertRef{tires}{slgf}
}
\keyword{datasets}
