% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torque.R
\docType{data}
\name{torque}
\alias{torque}
\title{Locknut data on torque required to tighten a fixture by plating method}
\format{
A data frame with 60 rows and 3 variables:
\describe{
  \item{Torque}{the response, the torque required to tighten the fixture.}
  \item{Fixture}{the type of fixture, bolt or mandrel.}
  \item{Plating}{the plating treatment, CW, HT, or PO.}
}
}
\usage{
data(torque)
}
\description{
Meek and Ozgur (1991) analyzes the torque required to strengthen a
fixture (bolt or mandrel) as a function of the fixture's plating method
(cadmium and wax, heat treating, and phosphate and oil, denoted CW, HT,
and PO, respectively).
}
\references{
\insertRef{torque}{slgf}
}
\keyword{datasets}
